+ type alias for sl dataset sample; fix imports; - split_dataset
