+++ src/utils/types.py
@@ -1,0 +1,3 @@
+from typing import Union, List
+
+SlDatasetSample = Union[List[List[Union[str, int]]], int]
